Before changing from 1 meter.

+++ point.hpp
@@ -13,11 +13,11 @@ public:
 		z = 0.0; 
 		rgb = 0.0; 
 	}
-	Point(float x2, float y2, float z2, float rgb2) : x(x2), y(y2), z(z2), rgb(rgb2) {}
-	float x; 
-	float y; 
-	float z; 
-	float rgb; 
+	Point(double x2, double y2, double z2, double rgb2) : x(x2), y(y2), z(z2), rgb(rgb2) {}
+	double x; 
+	double y; 
+	double z; 
+	double rgb; 
 
 	void draw(std::ostream & os){
 	 	os << "(" << x << ", " << y << ", " << z << ")" << endl;
